proper merging of existing cell values
in the previous implementation it was inconsistent how the merging
of existing cell values (for example if you built a spreadsheet, then
inserted some data then recompiled) the merge strategies for the various
file formats were different.  Google Sheets carried over the data while
CSV and Excel overwrote it.

now they all behave consistently and we could even make a CLI flag to
toggle whether we overwrite cell values or not