get rid of builtin functions & variables
I'm not removing the infrastructure needed to make them work, but the
functions themselves never really proved valuable.  (you could do
everything they were doing with native excel functions).

And if we're going to support module-loading of functions then that
makes even less of a case for builtins because we can use that facility