module loader wip: compiler_error(), eval variables
compiler_error() is a wrapper around panic!() that can be used whenever
the compiler finds itself in an inconsistent state and cannot recover.
in other words, one of our internal invariants has been violated. And we
wrap panic!() rather than returning a `Result` because it would litter
the call stack for cases that the caller can't do anything about anyway